调整窗口局部属性配置和全局属性配置的组合机制。调整默认字体属性设置。 调整渐变画刷实现。 ver: 4.5.0.0

+++ SOUI/src/helper/STaskHandler.cpp
@@ -23,7 +23,7 @@ void STaskHandler::start(const char *pszName, Priority priority)
         return;
     TCHAR szBuf[101];
 #ifdef _UNICODE
-    int nLen = MultiByteToWideChar(CP_ACP, 0, pszName, strlen(pszName), szBuf, 100);
+    int nLen = MultiByteToWideChar(CP_ACP, 0, pszName, (int)strlen(pszName), szBuf, 100);
     szBuf[nLen] = 0;
 #else
     strcpy_s(szBuf, 100, pszName);
